Crate Javadoc_to_Markdown

Re-exports

extern crate std;

extern crate clap;

extern crate regex;

extern crate threadpool;

use std::prelude::v1::*;

use clap::App;

use clap::Arg;

use std::fs;

use std::fs::File;

use std::io::prelude::*;

use std::path::Path;

use std::path::PathBuf;

use std::sync::Arc;

use threadpool::ThreadPool;

use model::model::Class;

use model::model::LineType;

use parse::parse::parse_file;

Modules

model

parse

Functions

document
Handles the thread pooling the application

find_java_files
Traverses the file structure to find all java files for parsing.

generate_markdown
Generates a markdown file for a java file Uses a Class struct to write the markdown

is_java_file

main